@JayCesar | refactor: moves classes and adds a logic to alow both Vendor and Reseller to have the same IDs of books
